requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.26.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.60.0
evaluate>=0.4.0
PyQt5>=5.15.0
safetensors>=0.3.0
 